pyproject.toml
[project]
name = "shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomisc>=17.7.7",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "dishka>=1.6.0",
    "fastapi[all,standard]>=0.115.12",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "ruff>=0.11.13",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]
